refactor: restructure test files for better data type testing workflow
- Extract comprehensive data type tests into `integration_datatypes_test.go`
- Rename `types_bench_test.go` → `datatypes_bench_test.go`
- Rename `benchmark_test.go` → `datatypes_conversion_bench_test.go`
- Create `pgarrow_test.go` with shared test helpers to eliminate duplication
- Remove duplicate helper functions from `integration_test.go`
- Add shared benchmark constants to prevent redeclaration errors

This restructuring makes data type issues easier to work with by isolating
all PostgreSQL data type tests in a single focused file, while maintaining
comprehensive test coverage and eliminating code duplication.
